+++ generators/app/templates/es/.eslintrc.js
@@ -47,7 +47,6 @@ module.exports = {
   },
   plugins: ['react', 'import', 'jsx-a11y', 'react-hooks', 'prettier'],
   rules: {
-    'arrow-parens': [ERROR, 'as-needed'],
     'class-methods-use-this': OFF,
     'comma-dangle': [ERROR, 'never'],
     'comma-spacing': [ERROR, { 'before': false, 'after': true }],