+++ generators/app/templates/ts/scripts/stylelint.ts
@@ -6,7 +6,7 @@ import stylelint from 'stylelint';
 import chalk from 'chalk';
 import ora from 'ora';
 
-process.on('unhandledRejection', err => {
+process.on('unhandledRejection', (err) => {
   throw err;
 });