+++ generators/app/templates/ts/scripts/test.ts
@@ -7,7 +7,7 @@ import * as jest from 'jest';
 process.env.BABEL_ENV = 'test';
 process.env.NODE_ENV = 'test';
 
-process.on('unhandledRejection', err => {
+process.on('unhandledRejection', (err) => {
   throw err;
 });