+++ generators/app/templates/ts/src/components/app/App.tsx
@@ -9,7 +9,7 @@ const App: React.FC = () => (
   <>
     <Header />
     <Switch>
-      {routes.map(route => (
+      {routes.map((route) => (
         <Route exact key={route.path} {...route} />
       ))}
       <Route component={Missing} />