+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 
 * Enable strict features in TypeScript template. _(dzervoudakes)_
 * Update all template dependencies. _(dzervoudakes)_
+* Reformat all scripts after an update to `Prettier`. _(dzervoudakes)_
 * Implement `postcss-preset-env` and uninstall `autoprefixer` plugin. _(dzervoudakes)_
-* Remove custom `react/sort-comp` lint rule. _(dzervoudakes)_
+* Remove `react/sort-comp` lint rule. _(dzervoudakes)_
 * Remove `no-undef` lint rule. _(dzervoudakes)_
+* Remove `arrow-parens` lint rule. _(dzervoudakes)_
 
 ## v1.0.3